requirements.txt
Flask==1.1.4
markupsafe==2.0.1
Flask-API==2.0
lxml==4.6.4
# viaa-chassis==0.1.3
uWSGI==2.0.18
requests>=2.23.0
pyjwt>=1.7.0
webvtt-py>=0.4.6
Werkzeug==1.0.0
certifi==2021.10.8
onelogin>=2.0.2
rdflib>=6.0.2
python3-saml>=1.12.0
flask-login>=0.5.0
MarkupSafe>=2.0.0
pandas>=1.3.5
argparse==1.4.0
skosify==2.3.0
SPARQLWrapper==1.8.5
markdown2>=2.4.2
